+++ game/scrabbleGame.py
@@ -6,16 +6,25 @@ from game.dictionary import(
     is_in_dictionary
 )
 
+class InvalidWordException(Exception):#NEW
+    pass
+class InvalidPlaceWordException(Exception):#NEW
+    pass
+class InvalidWordNoLetters(Exception):
+    pass
+
+
+
 class ScrabbleGame():
     def __init__(self, total_players: int):
         self.board = Board()
         self.bag_tiles = BagTiles()
         self.players:list[Player] = []
         for index in range(total_players):
-            self.players.append(Player(id=index, bag=self.bag_tiles))
+            self.players.append(Player(id=index, bag_tiles = self.bag_tiles))
         self.current_player = None
     
-    def next_turn(self):
+    def next_turn(self): 
         if self.current_player == None:
             self.current_player = self.players[0]
         else:            
@@ -24,18 +33,16 @@ class ScrabbleGame():
                 self.current_player = self.players[0]
             else:
                 self.current_player = self.players[index]
-    
-    def validate_word(self, player, word, location, orientation):
-        player_tiles = player.playertiles
-        tiles = [tile[0] for tile in player_tiles]
-        for letter in word:
-            if letter not in tiles:
-                return False
-        if not self.board.word_is_valid(word, location, orientation):
-            return False
+
+    def validate_word(self, word, location, orientation): #NEW #FIXING
+        if not Player().has_letters(word):
+            raise InvalidWordNoLetters("No cuenta con las fichas para armar esta palabra")
         if not is_in_dictionary(word):
-            return False
-        return True
+            raise InvalidWordException("Su palabra no existe en el diccionario")
+        if not self.board.valid_word_in_board(word, location, orientation):
+            raise InvalidPlaceWordException("Su palabra excede el tablero")
+        if not self.board.validate_word_place_board(word, location, orientation):
+            raise InvalidPlaceWordException("Su palabra esta mal puesta en el tablero")
                 
 """    
     def play():