+++ test/test_board.py
@@ -15,7 +15,7 @@ class TestBoard(unittest.TestCase):
         word = "Facultad"
         location = (5, 4)
         orientation = "H"
-        word_is_valid = board.word_is_valid(word, location, orientation)
+        word_is_valid = board.valid_word_in_board(word, location, orientation)
         assert word_is_valid == True
     
 
@@ -24,7 +24,7 @@ class TestBoard(unittest.TestCase):
         word = "Facultad"
         location = (14, 4)
         orientation = "H"
-        word_is_valid = board.word_is_valid(word, location, orientation)
+        word_is_valid = board.valid_word_in_board(word, location, orientation)
         assert word_is_valid == False
     
     def test_word_v_inside_board(self):
@@ -32,7 +32,7 @@ class TestBoard(unittest.TestCase):
         word = "Facultad"
         location = (5, 3)
         orientation = "V"
-        word_is_valid = board.word_is_valid(word, location, orientation)
+        word_is_valid = board.valid_word_in_board(word, location, orientation)
         assert word_is_valid == True
     
     def test_word_v_out_of_board(self):
@@ -40,7 +40,7 @@ class TestBoard(unittest.TestCase):
         word = "Facultad"
         location = (5, 14)
         orientation = "V"
-        word_is_valid = board.word_is_valid(word, location, orientation)
+        word_is_valid = board.valid_word_in_board(word, location, orientation)
         assert word_is_valid == False
     
     def test_board_is_empty(self):